requirements.txt
python-3.9.18
dash==2.11.1
dash_daq==0.4.0
dash_latex==0.1.1
gunicorn==20.0.4
numpy==1.25.0
pandas==2.1.3
plotly==5.5.0
pyDOE==0.3.8
scikit_learn==1.3.2
tensorflow-cpu==2.11.0
